refactor(Vsix): match namespace to "Json/SuggestedActions" folder.
