+++ src/LibraryManager.Vsix/Json/SuggestedActions/SuggestedActionProvider.cs
@@ -15,7 +15,7 @@ using Microsoft.WebTools.Languages.Json.Parser.Nodes;
 using Microsoft.WebTools.Languages.Json.VS.SuggestedActions;
 using Microsoft.WebTools.Languages.Shared.Parser.Nodes;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Json.SuggestedActions
 {
     [Export(typeof(IJsonSuggestedActionProvider))]
     [Name(nameof(SuggestedActionProvider))]
