+++ src/LibraryManager.Vsix/Json/SuggestedActions/UpdateSuggestedAction.cs
@@ -11,7 +11,7 @@ using Microsoft.WebTools.Languages.Shared.Editor.SuggestedActions;
 using Microsoft.WebTools.Languages.Shared.Parser.Nodes;
 using Microsoft.WebTools.Languages.Shared.Utility;
 
-namespace Microsoft.Web.LibraryManager.Vsix
+namespace Microsoft.Web.LibraryManager.Vsix.Json.SuggestedActions
 {
     internal class UpdateSuggestedAction : SuggestedActionBase
     {
